feat(lib): add helpers for all translations
