+++ src/lib/types.ts
@@ -98,6 +98,9 @@ export type Path = {
 export enum PathType {
   text = 'text',
   structured_text = 'structured_text',
+  structured_text_block = 'structured_text_block',
+  structured_text_inline_item = 'structured_text_inline_item',
+  structured_text_code = 'structured_text_code',
   media = 'media',
   id = 'id',
   number = 'number',
